Refactor virtualenv management
This PR attempts to simplify various aspects of virtualenv usage, as
well as some improvements to dependency tooling.

It based on applying techniques to the justfile that work well in
Makefiles.

Note: it uses a recently added feature (subsequent dependencies) so
requires just 0.9.9 or later.

1) Dependencies

We use a dependency heirarchy to split out the independant sections
of work. This allows us to manage prod/dev dependencies more easily,
and makes it easier for things to be automatically updated when
python dependencies change. As just always runs dependencies, each one
takes care to detect if it needs to run or not, in order to make things
fast in the normal case.

The developer now no longer needs to manually update their venv's
dependencies, they will automatically be updated for them if any of the
`requirements.*` files change. This is useful when switching between
branches. The developer can always refresh manually with `just devenv`
or `just prodenv`, of course.

Unlike commands, which can accept arguments, a dependency cannot be
parameterised, so we have to hardcode prod/dev versions of relavant
dependencies (i.e. `requirements-prod`, `prodenv`, `requirements-dev`,
and `devenv`). But dev/prod split is all we envision for this, and
encoding that division seems reasonable to help provoke careful thinking
about the difference.

2) Explicit paths

We remove the need to activate the virtualenv within the justfile, by
using explicit paths to the virtualenv's python and other executables.
This simplifies things further, and removes the need for sourcing
`setup_functions` all together.

3) Requirements files

With the dependencies split above, we can simply use
`--constraint requirements.prod.txt` within `requirements.dev.in` in
order to cleanly separate dev/prod requirements while still keeping them
in sync.

Previously, we had `-c requirement.prod.in`, which is effectively
a no-op as we don't pin versions in `requirements.prod.in`, and `-r
requirements.prod.in` which allowed mis-matched versions of packages
between dev and prod.

This change means you need to install both `requirements.prod.txt` and
`requirements.dev.txt`, however this is taken care of for the developer
automatically.

We also add --allow-unsafe to the pip-compile commands. This is going to
become the default when using --generate-hashes soon. And it allows us
to declare pip-tools as a formal dependency.

4) Easier updating

This change adds a new command `update`, for bumping python
dependencies. E.g

    just upgrade dev black      # upgrade black version in dev
    just upgrade dev            # upgrade all dev dependencies
    just upgrade prod requests  # upgrade requests in prod
    just upgrade prod           # upgrade all prod dependencies

This command can make good use of the other refactoring, and makes it
easier to keep things up to date, which is a Really Good Idea.