Pin `pip` to 25.0.1
This fixes our update tooling.

As the comment states, the latest version of `pip` is incompatible with
the latest released version of `pip-tools`:

https://github.com/jazzband/pip-tools/issues/2176

To unpin this, we need a fixed version of `pip-tools` to be available.

`pip` is pinned in two places, because there are two different entry
points to run updates:

* in the `justfile`, because the CI workflow that updates dependencies
  uses the `just virtualenv` recipe to install `pip` and `pip-tools`
* in the `requirements.dev.in`, because local development environments
  set up or updated with `just devenv` will get the version of `pip` in
  `requirements.dev.txt`, and the version in `requirements.dev.txt`
  would get updated without pinning there

(Note that it does seem reasonable that the CI uses the base `just
virtualenv` since it doesn't need all the development packages. We could
consider consolidating these routes, though we've talked about using `uv`
instead of `pip-tools`: `uv` likely makes simplifying this easy, because
the only requirement for updating the packages is the single binary
download of `uv`.)