Combine `requirements-dev` and `-prod` into one
Combine `requirements-dev` and `requirements-prod` into a single
`requirements` command, and replace `_compile` with `_uv`.

`_uv` respects the existing global and package-specific timestamps in
the lockfile.

We ignore the `UV_EXCLUDE_NEWER` environment variable if
it is set as we expect the user to be directly using `uv` commands if
they are setting that variable. A warning message is printed.