Disallow moving global timestamp backward in time
Make it impossible to set the global timestamp to a value that is older
than the global timestamp in the lockfile to prevent accidental
downgrades.
Print messages to the console if the global timestamp is being set to
anything other than the existing timestamp in the lockfile, or if the
user-specified timestamp is being deliberately ignored.
Package-specific timestamps can still be moved freely via `just upgrade`
as the user likely already knows what they are doing.