Enable `uv` on CI (#305)
We have released
[v1.6.1 of setup-action](https://github.com/opensafely-core/setup-action/releases/tag/v1.6.1),
which includes an `install-uv` option.
Set this to true in the template, as we assume most python projects
would like to use `uv` to manage dependencies.
We take advantage of the caching competencies of astral's setup-uv
action for speedier CI runs, by setting `setup-action`'s `cache` param
to "uv".