Add `_format-date` recipe for MacOS compatibility
We use the `date` command to parse relative dates like "7 days ago"
when passed as an argument to `just update-dependencies` etc.

On Linux, the default `date` command is from GNU coreutils, which can
take arguments like "7 days ago" or "yesterday" via the `--date` flag.
The BSD `date` command (default on MacOS) instead takes arguments like
"-7d" or "-1d" via the `-v` flag.

To support both, add a `_format-date` private recipe that checks which
`date` command is available and sets the appropriate flag. It is up to
the user to pass the correct format for their platform if they would
like to use relative dates; otherwise they should be pass
YYYY-MM-DD or YYYY-MM-DDTHH:MM:SSZ, which will be passed directly to
`uv`.