Rewrite `upgrade`
Most of the time `just upgrade` will look like `just requirements`. But,
if a package was previously pinned in `pyproject.toml` and is now
unpinned, `upgrade` will upgrade it to the latest version as of the
lockfile's timestamp cutoff, whereas `requirements` will not upgrade it.
This is a niche scenario, but probably justifies having both commands.

But since the above scenario is niche, we should suggest using
`update-dependencies` (to be rewritten in the next commit) instead of
`upgrade`. It is much more likely that someone is thinking to set a
new global timestamp cutoff and include newly available upgrades, which
is what `update-dependencies` does.