Access an `.env` file via `just _dotenv`
Provide a `dotenv-sample`, and add `.env` to `.gitignore`, such that
`just _dotenv` will load environment variables from `.env` if it exists.

This is essential for environment variables to be picked up in the
docker containers if they exist, as we will be configuring the
docker `just` recipes to use `just _dotenv`.