Add content to the Dockerfile
- Note that we use `.python-version` to specify the python version
instead of hardcoding it in the Dockerfile or in dependencies.txt,
to avoid the python versions being out of sync. This means that
`.python-version` needs to be removed from `.dockerignore`.

- Use a multi-stage process to ensure that compilers are not present in
the final project images.

- Derive `project-prod` and `project-dev` separately from `project-base`
since we only want to `COPY` the source code into the `-prod` image, but
not the `-dev` image. This is to avoid rebuilding every time the source
code changes during development, and to avoid accidentally using
embedded code rather than mounted code during development.