Add example `Dockerfile` that uses `uv`  (#308)
* Move template Dockerfile into `docker` folder

* Add content to the Dockerfile

- Note that we use `.python-version` to specify the python version
instead of hardcoding it in the Dockerfile or in dependencies.txt,
to avoid the python versions being out of sync. This means that
`.python-version` needs to be removed from `.dockerignore`.

- Use a multi-stage process to ensure that compilers are not present in
the final project images.

- Derive `project-prod` and `project-dev` separately from `project-base`
since we only want to `COPY` the source code into the `-prod` image, but
not the `-dev` image. This is to avoid rebuilding every time the source
code changes during development, and to avoid accidentally using
embedded code rather than mounted code during development.

* Add files for ubuntu dependencies

- `dependencies.txt` is empty; note that the `base-python` image already
contains instructions to install Python based on `.python-version`.

- `dependencies-build.txt` contains a commented-out placeholder for
`python<version>-dev`, since it might be possible that we would need to
build Python packages from scratch.

* Add entrypoint scripts

These scripts currently do nothing, but can be modified to pre-run
commands if needed in the future.

* Access an `.env` file via `just _dotenv`

Provide a `dotenv-sample`, and add `.env` to `.gitignore`, such that
`just _dotenv` will load environment variables from `.env` if it exists.

This is essential for environment variables to be picked up in the
docker containers if they exist, as we will be configuring the
docker `just` recipes to use `just _dotenv`.

* Add `docker-compose.yml` and docker `justfile`

Provide template `just` recipes for building and running docker
containers using `docker-compose`.

Note that to run `just docker/test`, a `.test.env` file must be
present in the base directory. We should `.gitignore` this
though.

* Update docker-related instructions in the README

We now have a whole `docker` folder, not just a `Dockerfile`.
Note that the other parts of the README will be updated separately.