Switch to v2 of update-dependencies-action
v2 of `update-dependencies-action` includes a 7-day cooldown period
in the default update command.
The cooldown period should be respected for dependency updates as
per our Dependency Updates Policy.
I have chosen to switch to v2 of update-dependencies-action instead
of keeping v1 and overriding the update command as this is the
template for future new repos in the organisation, so it feels a
bit wrong to use something other than the default input for a GHA
that is owned by us.