Implement dependency cooldown using `uv`
- Add `exclude-newer` timestamps to the `pyproject.toml` so that `uv`
will automatically track them for us.

- Rewrite `just check` such that it verifies that `uv.lock` is
reproducible from `pyproject.toml`. i.e. if either file is modified
without modifying the other via a `uv lock` command, then `check` will
fail. Note that the lockfile check must be run before `format` and
`lint` as the latter commands use `uv run` which automatically
overwrites the lockfile if it is out of date.

- Add a `bump-uv-cutoff` recipe, which allows one to update the
`exclude-newer` timestamp in the `pyproject.toml` programmatically.
By default, the new timestamp is set to midnight UTC 7 days from the
current date, but an alternative number of offset days can be provided.
(E.g. `just bump-uv-cutoff 0` will set the timestamp to midnight UTC
today.) The `tomlkit` package is used to parse and update the
`pyproject.toml` as it preserves formatting and comments.

- Add documentation in comments and in `DEVELOPERS.md`. This is to keep
the documentation close to the code.