Set dotenv-load in justfile and add _checkenv recipe
Most repos want .env loading, and the _dotenv recipe implies it.

Also adds a _checkenv recipe with comments as a reminder that
a .env created during just devenv or _dotenv will not be reloaded.
_checkenv can be used as a dependency for any recipes that require
a .env to be present, and will exit and prompt the user to run
just devenv first. This is a bit nicer than using _dotenv/devenv as
a dependency and having your just test/run etc fail on the first
run even though it created the required .env.