Ensure that tomlkit is using the correct version of python
Ensure that the bump-uv-cutoff recipe, which uses tomlkit, is
using a sufficiently modern version of python. Otherwise it may
fail in CI in unexpected ways when only the .python-version python
is available.

The datetime code used (datetime.UTC and proper isoformat handling)
is only available from 3.11. For this recipe, we don't care about
being consistent with the .python-version, it's only calculating
a timestamp to update in the pyproject.toml