Refactor the way we configure docker
Previouisly, we had place holders for versions of ubuntu/uv that needed
manually changing when the template is first used.

However, this meant that our docker set up was not directly testable,
w/o making those edits.

This change takes those placeholders and puts them in docker/.env with
some default values, passes them through as build args, which the docker
file uses.

This means that `just docker/build` works OOTB in the repo template, and
we can use it to test our *template* works as expected.

It's also a better UX, and defines a single point for configuration of
the docker image.

We also fix a few other small nits in the process

1) Which deadsnakes ppa name to use is automatically derived from the
   ubuntu version. Previously, you could mis-match these easily.

2) The uv version is defined once and resused, rather than twice (where
   they could technically be out of sync)

3) The tests dependend on a non-existant .test.env file, so commented
   that out