Just mount files needed, not whole of app
I had mistaken advised Alice to mount . into /app, but this was
a mistake, and meant that any change in . would invalidate the cache and
force a rebuild.

Now, we just explicitly mount the files needed to run the operation.
This means we get maxium cache efficiency, and also do not leave
a pointless layer via COPY