Add a functional test for `just upgrade-all`
We hadn't figured out how to test this yet, but now we have matching
uvmirror.txt, we need to make sure it works.

We create a fixture that provides a file based PEP 503 simple index that
we and manipulate and point uv at to test upgrade logic.  It sets it up
by populating the index with the currently locked package metadata,
which means that upgrade-all is a functional no-op.

We can then play around with manipulating the versions and upload
timestamps and assert that our delayed upgrade logic works as expected.